ALPRP/requirements.txt
flask
ffmpeg-python
xmltodict
pytesseract
opencv-python
matplotlib
imutils
scikit-image
scikit-learn
pandas
numpy